+++ packages/gatsby-plugin-sharp/src/index.js
@@ -40,10 +40,7 @@ exports.setBoundActionCreators = actions => {
   boundActionCreators = actions
 }
 
-// We set the queue to a Map instead of an array to easily search in onCreateDevServer Api hook
-const queue = new Map()
-exports.queue = queue
-
+const cachedOutputFiles = new Map()
 function queueImageResizing({ file, args = {}, reporter }) {
   const pluginOptions = getPluginOptions()
   const options = healOptions(pluginOptions, args, file.extension)
@@ -107,21 +104,34 @@ function queueImageResizing({ file, args = {}, reporter }) {
     outputPath: outputFilePath,
   }
 
-  queue.set(prefixedSrc, job)
+  const outputFile = path.join(job.outputDir, job.outputPath)
+  let finishedPromise = Promise.resolve()
 
-  // schedule job immediately - this will be changed when image processing on demand is implemented
-  const finishedPromise = scheduleJob(
-    job,
-    boundActionCreators,
-    pluginOptions,
-    reporter
-  ).then(() => {
-    queue.delete(prefixedSrc)
-  })
+  if (cachedOutputFiles.has(outputFile)) {
+    finishedPromise = cachedOutputFiles.get(outputFile)
+  }
+
+  // Check if the output file already exists or already is being created.
+  // TODO: Remove this when jobs api is stable, it will have a better check
+  if (!fs.existsSync(outputFile) && !cachedOutputFiles.has(outputFile)) {
+    // schedule job immediately - this will be changed when image processing on demand is implemented
+    finishedPromise = scheduleJob(
+      job,
+      boundActionCreators,
+      pluginOptions,
+      reporter
+    ).then(res => {
+      cachedOutputFiles.delete(outputFile)
+
+      return res
+    })
+
+    cachedOutputFiles.set(outputFile, finishedPromise)
+  }
 
   return {
     src: prefixedSrc,
-    absolutePath: path.join(outputDir, outputFilePath),
+    absolutePath: outputFile,
     width,
     height,
     aspectRatio,