+++ packages/gatsby-plugin-google-analytics/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.29](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-google-analytics@2.1.28...gatsby-plugin-google-analytics@2.1.29) (2019-11-26)
+
+### Bug Fixes
+
+- **gatsby-plugin-google-analytics:** Update typings ([#19684](https://github.com/gatsbyjs/gatsby/issues/19684)) ([524862a](https://github.com/gatsbyjs/gatsby/commit/524862a))
+
+## [2.1.28](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-google-analytics@2.1.27...gatsby-plugin-google-analytics@2.1.28) (2019-11-18)
+
+### Bug Fixes
+
+- **gatsby-plugin-google-analytics:** Don't create tracking cod… ([#19592](https://github.com/gatsbyjs/gatsby/issues/19592)) ([e883487](https://github.com/gatsbyjs/gatsby/commit/e883487))
+
 ## [2.1.27](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-google-analytics@2.1.26...gatsby-plugin-google-analytics@2.1.27) (2019-11-15)
 
 **Note:** Version bump only for package gatsby-plugin-google-analytics