+++ packages/gatsby-plugin-typography/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.18](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typography@2.3.17...gatsby-plugin-typography@2.3.18) (2019-11-26)
+
+**Note:** Version bump only for package gatsby-plugin-typography
+
 ## [2.3.17](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typography@2.3.16...gatsby-plugin-typography@2.3.17) (2019-11-15)
 
 **Note:** Version bump only for package gatsby-plugin-typography