+++ packages/gatsby-react-router-scroll/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.17](https://github.com/gatsbyjs/gatsby/compare/gatsby-react-router-scroll@2.1.16...gatsby-react-router-scroll@2.1.17) (2019-11-26)
+
+**Note:** Version bump only for package gatsby-react-router-scroll
+
 ## [2.1.16](https://github.com/gatsbyjs/gatsby/compare/gatsby-react-router-scroll@2.1.15...gatsby-react-router-scroll@2.1.16) (2019-11-15)
 
 **Note:** Version bump only for package gatsby-react-router-scroll